only copy non dot files from target repo

+++ src/cmd/up.rs
@@ -1,7 +1,7 @@
 use regex::Regex;
 use serde_derive::Deserialize;
 use std::env;
-use std::fs::{read_dir, remove_file, rename, symlink_metadata, File};
+use std::fs::{read_dir, read_link, remove_file, rename, symlink_metadata, File};
 use std::io::prelude::*;
 use std::os::unix::fs::symlink;
 use std::path::{Path, PathBuf};
@@ -25,9 +25,10 @@ pub fn link_dotfiles() -> std::io::Result<()> {
     let mut dotrc_path: PathBuf = PathBuf::new();
 
     if !dotrc_env_var_set {
-        // DOTRC env var not set
-        dotrc_path.push(&format!("{}/.dot.toml", out_dir));
+        // DOTRC env var not set, use dot.toml from $HOME
+        dotrc_path.push(&format!("{}/.dotrc", out_dir));
     } else {
+        // DOTRC env var set, use env var path
         dotrc_path.push(dotrc_env_var.unwrap());
     }
 
@@ -46,6 +47,12 @@ pub fn link_dotfiles() -> std::io::Result<()> {
 
         for path in paths {
             let path_str = path.unwrap().path().display().to_string();
+            let filename = Path::new(&path_str).file_name().unwrap().to_str().unwrap();
+
+            if filename.chars().nth(0).unwrap() == '.' {
+                continue;
+            }
+
             let mut should_exclude: bool = false;
             for exclude in &config.excludes {
                 let re = Regex::new(&exclude).unwrap();
@@ -57,12 +64,15 @@ pub fn link_dotfiles() -> std::io::Result<()> {
             if should_exclude {
                 continue;
             }
-            let filename = Path::new(&path_str).file_name().unwrap().to_str().unwrap();
+
             let out_path = &format!("{}/.{}", out_dir, filename);
             if Path::new(&out_path).exists() {
                 let metadata = symlink_metadata(&out_path)?;
                 let file_type = metadata.file_type();
                 if file_type.is_symlink() {
+                    if read_link(&out_path).unwrap().to_str().unwrap() == path_str {
+                        continue;
+                    }
                     remove_file(&out_path)?;
                 }
                 if file_type.is_file() {
